fixed request fee bug (int to float)

+++ server/src/models/Event.js
@@ -7,7 +7,7 @@ const Event = sequelize.define('Event', {
   eventImage: { type: DataTypes.STRING }, // Store image path
   eventDateTime: { type: DataTypes.DATE, allowNull: false },
   eventLocation: { type: DataTypes.STRING, allowNull: false },
-  requestFee: { type: DataTypes.INTEGER, allowNull: false },
+  requestFee: { type: DataTypes.FLOAT, allowNull: false },
   DJID: { type: DataTypes.INTEGER, allowNull: false }, // Foreign key
 }, { timestamps: false });
